// src/features/RegistrationUser/model/consts/consts.ts
 export enum ValidateRegistrationErrors {
    NO_DATA = 'NO_DATA',
    SERVER_ERROR = 'SERVER_ERROR',
    PASSWORD_MISSMATCH = 'PASSWORD_MISSMATCH',
    SHORT_PASSWORD = 'SHORT_PASSWORD',
    INCORRECT_USERNAME = 'INCORRECT_USERNAME',
    INCORRECT_LASTNAME = 'INCORRECT_LASTNAME',
    INCORRECT_FIRSTNAME = 'INCORRECT_FIRSTNAME',
    INCORRECT_AGE = 'INCORRECT_AGE',
    INCORRECT_CITY = 'INCORRECT_CITY'
 }